rename vfs name file
