+++ dev/connection_holder.h
@@ -8,7 +8,7 @@
 #endif
 
 #include "error_code.h"
-#include "vfs_object.h"
+#include "vfs_name.h"
 #include "db_open_mode.h"
 
 namespace sqlite_orm {
